fix: agent enrollment, WebSocket scheme, Windows v0.58 API, relay JWT auth
- Fix enrollment field name: send "token" instead of "enrollmentToken"
- Convert https:// to wss:// for relay WebSocket URL construction
- Fix Windows crate v0.58 API changes (DXGI Map, flag types, ConPTY,
  HANDLE constructor, unsafe Send impl)
- Disable turbojpeg SIMD for Windows cross-compilation
- Add missing Windows feature gates (Win32_Graphics_Gdi, Win32_System_IO)
- Fix ARM32 statvfs u64 casts for disk info
- Fix relay JWT secret mismatch: use shared getJwtSecret() instead of
  hardcoded fallback

+++ agent/crates/agent-linux/src/system_info.rs
@@ -242,9 +242,9 @@ fn parse_disk_info() -> Vec<DiskInfo> {
         }
 
         let block_size = stat.f_frsize as u64;
-        let total_bytes = stat.f_blocks * block_size;
-        let available_bytes = stat.f_bavail * block_size;
-        let free_bytes = stat.f_bfree * block_size;
+        let total_bytes = stat.f_blocks as u64 * block_size;
+        let available_bytes = stat.f_bavail as u64 * block_size;
+        let free_bytes = stat.f_bfree as u64 * block_size;
         let used_bytes = total_bytes.saturating_sub(free_bytes);
 
         // Skip zero-size filesystems
